capstone_backend/backend/requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.0.5
django-filter==2.4.0
django-phone-field==1.8.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
dnspython==2.1.0
mysqlclient==2.0.3
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.2.4
uuid==1.30
